debug_mem needs rewriting to work with delete[] on gcc. gcc build fails on throws. debug::xError class needs implementing, maybe? It gives good results on VS2010, and threading is off...
git-svn-id: file:///mnt/svn/rtmath@123 1d52797c-35fc-48e8-8e80-2db72aa187e2

+++ rtmath/rtmath/debug_mem.h
@@ -44,8 +44,8 @@ inline void* operator new(size_t size)
 	//AddTrack(p, size, file, line);
 	if (__caller__) fprintf(stderr,"\nHEAP: new called from: %s\n", __caller__);
 	if (__file__) fprintf(stderr, " file: %s\n", __file__);
-	if (__line__) fprintf(stderr, " line: %d\n", __line__);
-	fprintf(stderr, " size: %d\n", size);
+	if (__line__) fprintf(stderr, " line: %d\n", (int) __line__);
+	fprintf(stderr, " size: %d\n", (int) size);
 	int lengthInt = size / sizeof(int);
 	int *end = &(((int*) p)[lengthInt]);
 	fprintf(stderr, " memory range: %p - %p\n", p, end);
@@ -71,8 +71,8 @@ inline void* operator new[](size_t size)
 	//AddTrack(p, size, file, line);
 	if (__caller__) fprintf(stderr,"\nHEAP: new called from: %s\n", __caller__);
 	if (__file__) fprintf(stderr, " file: %s\n", __file__);
-	if (__line__) fprintf(stderr, " line: %d\n", __line__);
-	fprintf(stderr, " size: %d\n", size);
+	if (__line__) fprintf(stderr, " line: %d\n", (int) __line__);
+	fprintf(stderr, " size: %d\n", (int) size);
 	int lengthInt = size / sizeof(int);
 	int *end = &(((int*) p)[lengthInt]);
 	fprintf(stderr, " memory range: %p - %p\n", p, end);
